[#44] refactor : locationPage 이동 및 type분리
