+++ src/components/User/RegisterLocation/types.ts
@@ -1,0 +1,11 @@
+type Address = {
+  [key: string]: string;
+};
+export interface LocationResult {
+  address: Address;
+  address_name: string;
+  address_type: string;
+  road_address: null;
+  x: string;
+  y: string;
+}
